requirements.txt
pytesseract==0.3.13
Pillow==11.2.1
requests==2.32.3
python-dotenv==1.0.1
